bugfix where after ~1 min board would spam CAN messages

+++ Core/Src/power_board/power_board.c
@@ -4,7 +4,7 @@
 #include "main.h"
 
 
-uint16_t heartbeat_10000ms = 0;
+uint32_t heartbeat_10000ms = 0;
 bool kill_flag, voltage_request = false;
 uint16_t voltage_reading = 0;
 uint64_t TxMailbox[0];
